[Docs] Add clarification for target types in CrossEntropyLoss doc
CrossEntropyLoss function requires that target for class indices are provided
as a long and class probabilities are provided as a float datatype.

The CrossEntropyLoss function distinguish the two scenarios (indices and probabilities)
by comparing the shapes. When input and target shapes are the same it’s a case for
probabilities otherwise it will be used as a class index as already covered in the doc.
The related code is here,
https://github.com/pytorch/pytorch/blob/main/aten/src/ATen/native/LossNLL.cpp#L624

I think the current documentation is great but seems like it can confuse users about
types as reported in the issues so this PR adds a bit more clarification.

Fixes #137188

Signed-off-by: Sahdev Zala <spzala@us.ibm.com>